test(tui): move inline tests to sibling files
